undo module name change
